+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.1
 - Bug fix: Preview text would trigger a disabled sound
+- Preview sound will trigger (text, fx and art)
 
 # 0.1.0
 First release